fix: serialisation of `since` datetime (#88)
* fix: serialisation of `since` datetime

* The `since` parameter of many jobs is represented in Python with a
datetime.datetime object. This cannot be directly serialised into a str,
so a `default=str` was added to `json.dumps`. However, this serialises
the date into a non-standard format instead of ISO 8601 which is what
job handlers expect to receive.

* Added a custom default function to turn any datetime into an ISO 8601
formatted string, without affecting any other job args.

* fix: use same job arg serialisation everywhere

* To ensure the UI appears consistent with the actual job arguments, I
generalised the `json.dumps` custom `default` function and used it
everywhere a job is serialised.

* The uses in `services/results.py` are purely UI-related but it's still
important to ensure the timestamps are displayed consistently (as well
as any future changes to the custom serialiser)

* fix: add timezone info to `since` argument

* Some job implementations downstream assume the `since` argument in
jobs is an ISO timestamp *including timezone info*. We store all dates
in the DB in UTC but without associated TZ info.

* It is safe to add in the TZ info (UTC) when setting the `since` arg as
we know it is always in the UTC zone. Job implementors can then more
easily (and without ambiguity about the TZ) use the `since` arg.

* fix: behaviour of since argument

* Changed the `since` argument so task implementors always receive a
timestamp with the `+00:00` syntax instead of `Z` which is unsupported
by Python <3.10

* Added an error message shown on the admin dashboard when running a job
with an invalid `since` timestamp (i.e. not ISO compliant)

* Added unit test for the custom `json.dumps` serialiser that
stringifies dates to ISO timestamps

* Modified the job unit test to use the `since` argument, and added a
test case for an invalid timestamp to ensure the correct error message
is returned to the client

* fix: move validation logic back to schema

* Previously, the PredefinedArgsSchema was not actually being used for
validation due to a bug in the logic of `get_data_type`
(https://github.com/inveniosoftware/invenio-jobs/blob/ccf7ee3af0126c04ab8286878213674b66206bfa/invenio_jobs/services/schema.py#L136)

* This bug has been fixed, and the validation logic has been moved to
the schema, ensuring the proper separation of concerns between the job
argument builder and the request handler.

* The previous error object used at the job argument builder level has
been removed, as this error is now provided directly by Marshmallow.
This still ensures the provided `since` value (if applicable) is a valid
ISO timestamp.

* tests: add case for job run without provided arguments

* Ensures that the `custom` schema is correctly assigned when a job is
run without args