fix: datetime not being correctly serialised in scheduled jobs
* In https://github.com/inveniosoftware/invenio-jobs/pull/88 the
serialisation of the `since` argument was moved downstream; it was
originally the responsibility of the job implementor.

* The argument (and any other datetimes) was being serialised correctly
on manual job runs, but argument serialisation is handled by different
code for scheduled jobs. And I had forgotten to update this code to also
support the serialisation, leading to an error only on non-manual job
runs.

* Updated JobEntry to correctly serialise the job's default args through
`job_arg_json_dumper` as already done in models.py